+++ Projet Onssa/LrgForm.cs
@@ -47,9 +47,14 @@ namespace Projet_Onssa
 
                     ctx.LrgSet.Add(olrg);
                     ctx.SaveChanges();
-                    MessageBox.Show("ok");
+                    MessageBox.Show("Ajouté avec succès");
+                    DeclarationGlobale.vider(this);
                     cb_code.DataSource = ctx.LrgSet.ToList();
 
+                }
+                catch (FormatException)
+                {
+                    MessageBox.Show("Format text non valide !", "Erreur", MessageBoxButtons.OK, MessageBoxIcon.Error);
                 }
                 catch (Exception o)
                 {
@@ -81,50 +86,63 @@ namespace Projet_Onssa
 
         private void btn_Modifier_Click(object sender, EventArgs e)
         {
-            using (OnssaModelContainer4 ctx = new OnssaModelContainer4())
+            DialogResult r = MessageBox.Show("Êtes-vous sûr de vouloir Modifier ?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (r == DialogResult.Yes)
             {
-                try
+                using (OnssaModelContainer4 ctx = new OnssaModelContainer4())
                 {
-                    ctx.Entry(lr).State = System.Data.Entity.EntityState.Modified;
-                    Paragraphe Op = ctx.ParagrapheSet.Find(cb_numpara.SelectedValue);
+                    try
+                    {
+                        ctx.Entry(lr).State = System.Data.Entity.EntityState.Modified;
+                        Paragraphe Op = ctx.ParagrapheSet.Find(cb_numpara.SelectedValue);
 
-                    lr.NumLrg = cb_Num.Text.ToString();
-                    lr.InfoParagraphe = Op;
-                    lr.DescriptionLrg = txtarea_desc.Text;
+                        lr.NumLrg = cb_Num.Text.ToString();
+                        lr.InfoParagraphe = Op;
+                        lr.DescriptionLrg = txtarea_desc.Text;
 
-                    lr.CodeLrg = int.Parse(Op.NumPar.ToString() + cb_Num.Text.ToString());
+                        lr.CodeLrg = int.Parse(Op.NumPar.ToString() + cb_Num.Text.ToString());
 
-                    ctx.SaveChanges();
-                    MessageBox.Show("ok");
-                    cb_code.DataSource = ctx.LrgSet.ToList();
+                        ctx.SaveChanges();
+                        MessageBox.Show("Modifier avec succès");
+                        DeclarationGlobale.vider(this);
+                        cb_code.DataSource = ctx.LrgSet.ToList();
+                    }
+                    catch (FormatException)
+                    {
+                        MessageBox.Show("Format text non valide !", "Erreur", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    }
+                    catch (Exception o)
+                    {
+                        MessageBox.Show(o.Message);
+                    }
 
-                }
-                catch (Exception o)
-                {
-                    MessageBox.Show(o.Message);
                 }
             }
         }
 
         private void btn_Supprimer_Click(object sender, EventArgs e)
         {
-            using (OnssaModelContainer4 ctx = new OnssaModelContainer4())
+            DialogResult r = MessageBox.Show("Êtes-vous sûr de vouloir supprimer ?", "Confirmation", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (r == DialogResult.Yes)
             {
-                try
+                using (OnssaModelContainer4 ctx = new OnssaModelContainer4())
                 {
-                    ctx.Entry(lr).State = System.Data.Entity.EntityState.Modified;
-
-                    ctx.LrgSet.Remove(lr);
-                    ctx.SaveChanges();
-                    MessageBox.Show("Supp ok");
-                    cb_code.DataSource = ctx.LrgSet.ToList();
+                    try
+                    {
+                        ctx.Entry(lr).State = System.Data.Entity.EntityState.Modified;
+                        ctx.LrgSet.Remove(lr);
+                        ctx.SaveChanges();
+                        cb_code.DataSource = ctx.LrgSet.ToList();
+                        MessageBox.Show("Supprimer avec succès");
+                        DeclarationGlobale.vider(this);
+                    }
+                    catch (Exception o)
+                    {
+                        MessageBox.Show(o.Message);
+                    }
 
                 }
-                catch (Exception o)
-                {
-                    MessageBox.Show(o.Message);
-                }
-            }
+            }       
         }
     }
 }
